style: Moved all the types into the proper folder for addresses.
